# docs/swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/auth/employee/sign-up":{"post":{"operationId":"AuthController_employeeSignUp","summary":"Регистрация нового работника","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpDto"}}}},"responses":{"201":{"description":""}},"tags":["Авторизация"]}},"/api/auth/accountant/sign-up":{"post":{"operationId":"AuthController_accountantSignUp","summary":"Регистрация нового бухгалтера","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpDto"}}}},"responses":{"201":{"description":""}},"tags":["Авторизация"]}},"/api/auth/sign-in":{"post":{"operationId":"AuthController_signIn","summary":"Авторизация пользователя","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDto"}}}},"responses":{"200":{"description":""},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["Авторизация"]}},"/api/auth/logout":{"get":{"operationId":"AuthController_logout","parameters":[],"responses":{"200":{"description":""}},"tags":["Авторизация"],"security":[{"bearer":[]}]}},"/api/auth/refresh":{"get":{"operationId":"AuthController_refresh","parameters":[],"responses":{"200":{"description":""}},"tags":["Авторизация"]}},"/api/account":{"get":{"operationId":"AccountController_profile","summary":"Получение профиля авторизованного пользователя","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileDto"}}}}},"tags":["Аккаунт пользователя"],"security":[{"bearer":[]}]}}},"info":{"title":"FSP KUBOK","description":"","version":"0.1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"Bearer","bearerFormat":"JWT","type":"http","in":"Header"}},"schemas":{"SignUpDto":{"type":"object","properties":{"handle":{"type":"string"},"email":{"type":"string"},"lastName":{"type":"string"},"firstName":{"type":"string"},"middleName":{"type":"string"},"password":{"type":"string"}},"required":["handle","email","lastName","firstName","middleName","password"]},"SignInDto":{"type":"object","properties":{"login":{"type":"string"},"password":{"type":"string"}},"required":["login","password"]},"ErrorDto":{"type":"object","properties":{"message":{"type":"string"},"error":{"type":"string"},"statusCode":{"type":"number"}},"required":["message","error","statusCode"]},"ProfileDto":{"type":"object","properties":{"id":{"type":"string"},"handle":{"type":"string"},"email":{"type":"string"},"lastName":{"type":"string"},"firstName":{"type":"string"},"middleName":{"type":"string"},"role":{"type":"string"}},"required":["id","handle","email","lastName","firstName","middleName","role"]}}}}